Function sunrise_libutils::print_hexdump_as_if_at_addr

pub fn print_hexdump_as_if_at_addr<T: Write>(
    f: &mut T, 
    mem: &[u8], 
    display_addr: usize
)

Makes a hexdump of a slice, but display different addresses. Used for displaying memory areas which are not identity mapped in the current pages

This function follows the output format of the xxd command so you can easily diff between two hexdumps on the host machine, to help debugging sessions.